+++ test/src/test_tap_interface.cpp
@@ -4,9 +4,10 @@
 // Distributed under the "BSD License". See the accompanying LICENSE.rst file.
 
 #include <gtest/gtest.h>
+#include <platform/config.hpp>
 #include <tunnel/tap_interface.hpp>
-#include <unistd.h>
 
+#if defined(PLATFORM_LINUX)
 TEST(test_tap_interface, construct_no_su_fail)
 {
     tunnel::tap_interface t;
@@ -25,3 +26,5 @@ TEST(test_tap_interface, create_no_su_expect_fail)
                      std::system_error);
     }
 }
+
+#endif